Better use of bgr + Global
First, let-bind the correct base directory for bgr-spawn instead of
shoving a "cd" onto the beginning of the command. Also, detect the
presence of GTAGS and, if present, invoke "global -u", which is a lot
faster. Finally, do not use verbose output from "gtags" because it makes
it ten times slower (from 8 seconds to 100+ seconds).