Prevent double-linking of dirs
zsh is a clever girl. When asked to create a link from /path/to/dirname
with the name "dirname," it dutifully creates a link at
/path/to/dirname/dirname, assuming that you meant "put my link in this
directory," which I suppose a rational person would. This script may
attempt to link directories that already exist, though, so just check
that first and don't try to create the link if it's already there.