Use auto-fill and electric pairs in prog-modes
This prevents *losing* auto-fill in the Magit commit buffer because the
commit message area is not a comment; `comment-auto-fill-only-comments`
is set locally in modes derived from `prog-mode` only.

This also sets `electric-pair-local-mode` in all programming
modes (instead of turning it on when PHP Mode is loaded and leaving it
on globally, which made no sense).

Protip: Electric Pair Mode doesn't work very well in prose, e.g., Org.