Fix dumb bug in subtree arg to org skip funcs
I think this was copypasta from someone else's org skip function and
when I decided to make the let-var `end` instead of `subtree-end` I
prepended it instead of replacing it, so it was actually evaluating this
unbound variable and returning `nil` every time.

I don't really use the `subtree` arg, so I don't think it affected my
agenda, but it was wrong, and wrong things must be fixed.