Don't let `C-d` kill tmux windows anymore
Yeah so as it turns out, I have to press `C-d` to send EOF or to
scroll half a page down in vi-like programs (such as `less`) all the
time, so when that kills the current window it really harshes my
mellow.

This will still do the same thing if there is only one window, but
that's fairly rare. Until I find a better way (which might be "kill
this thing entirely and pay more attention, you idiot") I'm going to
leave it like this.