Skip scheduled items in "for meetings," show fortnight in review
If I schedule an item, it's because I don't want to see it until then,
so hide it from other places, mainly the "for meetings" section, which
already had such selective criteria that I didn't realize it was
showing scheduled items.

Additionally, show a fortnight worth of upcoming agenda days on the
agenda review so I can get a better picture of what's coming up.